user follow book follow

+++ Controller/User/User_Follow.js
@@ -1,45 +1,41 @@
 const User = require("../../models/User");
 const User_Follow = require("../../models/User_Follow");
 
-// follow user
-exports.follow = async (req, res, next) => {
+// follow toggle
+exports.toggle = async (req, res, next) => {
 	try {
-		FollowThisUser = await User.findOne({
-			where: { Username: req.body.FollowUsername },
+		const Follower = req.person;
+		const Followed = await User.findOne({
+			where: { Username: req.body.Username },
 		});
-
-		let already_followed = await User_Follow.findOne({
-			where: { UserID: req.person.ID, FollowedUserID: FollowThisUser.ID },
-		});
-		if (!already_followed && req.person.ID !== FollowThisUser.ID) {
-			let follow_user = new User_Follow({
-				UserID: req.person.ID,
-				FollowedUserID: FollowThisUser.ID,
+		if (Follower.ID !== Followed.ID) {
+			let already_followed = await User_Follow.findOne({
+				where: { UserID: Follower.ID, FollowedUserID: Followed.ID },
 			});
-			await follow_user.save();
-			res.status(200).send({ Response: req.body.FollowUsername + " Followed" });
-		} else {
-			res.status(400).send({ Response: "Error" });
-		}
-	} catch (e) {
-		res.status(400).send({ Response: "Error" });
-		console.log(e);
-	}
-};
+			var result = null;
+			//delete row
+			if (already_followed) {
+				const tmp = await User_Follow.destroy({
+					where: { UserID: Follower.ID, FollowedUserID: Followed.ID },
+				});
+				result = false;
+			}
+			//follow
+			else {
+				const tmp = new User_Follow({
+					UserID: Follower.ID,
+					FollowedUserID: Followed.ID,
+				});
+				await tmp.save();
+				result = true;
+			}
 
-// unfollow user
-exports.unfollow = async (req, res, next) => {
-	try {
-		UnfollowThisUser = await User.findOne({
-			where: { Username: req.body.UnfollowUsername },
-		});
-		let UnfollowedUser = await User_Follow.destroy({
-			where: { UserID: req.person.ID, FollowedUserID: UnfollowThisUser.ID },
-		});
-		if (UnfollowedUser) {
-			res
-				.status(200)
-				.send({ Response: req.body.UnfollowUsername + " Unfollowed" });
+			res.status(200).send({
+				Response: "Done",
+				User_ID: Followed.ID,
+				Username: Followed.Username,
+				Following_State: result,
+			});
 		} else {
 			res.status(400).send({ Response: "Error" });
 		}
